PythonService1/requirements.txt
Flask==3.0.3
opentelemetry-distro[otlp]

